requirements.txt
Django==5.1.4
djangorestframework==3.15.2
python_dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
 